+++ lib/constructs/CdkEcrDeploymentStackProps.ts
@@ -1,5 +1,5 @@
 import { NestedStackProps } from 'aws-cdk-lib';
-import { CohereRerankV3Services } from '../../lib/CohereRerankV3ServicesStackProps';
+import { CohereRerankV3Services } from '../CohereRerankV3ServicesStackProps';
 
 /**
  * Properties for CdkEcrDeploymentStack.
