+++ lib/constructs/cohere-embded-v3-app-runner.ts
