+++ lib/constructs/cohere-embed-v3-ecr.ts
