+++ lib/constructs/cohere-rerank-v3-vpc.ts
